[RG-n2ntraffic] Add get_graph_for_node_link
This new function implements the previous functionality that
views/api:link() had. It no longer uses 'Links' objects to
resolve a node link, but directly searches for the required
'Graph' object. Then, if the 'separate' argument is provided,
calls graph_for_each interface' to get all the separate graphs
that form this aggregation, otherwise just returns the png of
the aggregate graph using a reverse url resolution of 'get-png-data'.